Chat works good enough

+++ lab_8/simple_chat/backend/src/main/java/com/simplechat/backend/socket/SocketModule.java
@@ -9,9 +9,11 @@ import com.corundumstudio.socketio.listener.DisconnectListener;
 import com.simplechat.backend.constants.Constants;
 import com.simplechat.backend.model.Message;
 import com.simplechat.backend.model.TypingDTO;
+import com.simplechat.backend.service.RoomService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
+import java.util.List;
 import java.util.stream.Collectors;
 
 @Component
@@ -20,10 +22,12 @@ public class SocketModule {
 
     private final SocketIOServer server;
     private final SocketService socketService;
+    private final RoomService roomService;
 
-    public SocketModule(SocketIOServer server, SocketService socketService) {
+    public SocketModule(SocketIOServer server, SocketService socketService, RoomService roomService) {
         this.server = server;
         this.socketService = socketService;
+        this.roomService = roomService;
         server.addConnectListener(onConnected());
         server.addDisconnectListener(onDisconnected());
         server.addEventListener("send_message", Message.class, onChatReceived());
@@ -50,11 +54,17 @@ public class SocketModule {
             var params = client.getHandshakeData().getUrlParams();
             String room = params.get("room").stream().collect(Collectors.joining());
             String username = params.get("username").stream().collect(Collectors.joining());
+            List<String> usersInRoom = roomService.getUsersInRoom(room);
             client.joinRoom(room);
-            socketService.saveInfoMessage(client, String.format(Constants.WELCOME_MESSAGE, username), room);
-            log.info("Socket ID[{}] - room[{}] - username [{}]  Connected to chat module through", client.getSessionId().toString(), room, username);
-        };
 
+            if (!usersInRoom.contains(username)) {
+                socketService.saveInfoMessage(client, String.format(Constants.WELCOME_MESSAGE, username), room);
+                log.info("Socket ID[{}] - room[{}] - username [{}]  Connected to chat module through", client.getSessionId().toString(), room, username);
+            } else {
+                log.info("User {} already in room {}", username, room);
+            }
+            roomService.addUserToRoom(room, username);
+        };
     }
 
     private DisconnectListener onDisconnected() {
@@ -62,8 +72,15 @@ public class SocketModule {
             var params = client.getHandshakeData().getUrlParams();
             String room = params.get("room").stream().collect(Collectors.joining());
             String username = params.get("username").stream().collect(Collectors.joining());
-            socketService.saveInfoMessage(client, String.format(Constants.DISCONNECT_MESSAGE, username), room);
-            socketService.removeTypingUser(room, username);
+            roomService.removeUserFromRoom(room, username);
+
+            List<String> usersInRoom = roomService.getUsersInRoom(room);
+
+            if (!usersInRoom.contains(username)) {
+                log.info("Socket ID[{}] - room[{}] - username [{}]  discnnected to chat module through", client.getSessionId().toString(), room, username);
+                socketService.saveInfoMessage(client, String.format(Constants.DISCONNECT_MESSAGE, username), room);
+                socketService.removeTypingUser(room, username);
+            }
             log.info("Socket ID[{}] - room[{}] - username [{}]  discnnected to chat module through", client.getSessionId().toString(), room, username);
         };
     }
